requirements.txt

segmentation_models_pytorch>=0.4.0
torchstain[torch]>=1.4.1
torchsummary>=1.5.1
torchmetrics>=1.6.1
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
PyQt6>=6.4.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
alembic>=1.12.0
configparser>=6.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
pre-commit>=4.1.0
isort>=6.0.1
autopep8>=2.3.2